+++ test/OpenFeature.Contrib.Providers.Flagd.E2e.ProcessTest/BeforeHooks.cs
@@ -24,9 +24,15 @@ public class BeforeHooks
         var scenarioTags = scenarioInfo.Tags;
         var featureTags = featureInfo.Tags;
         var tags = new HashSet<string>(scenarioTags.Concat(featureTags));
+<<<<<<< add-flagd-config-e2e-tests
         Skip.If(!tags.Contains("in-process"), "Skipping scenario because it is not for the in-process resolver.");
 
         // TODO: https://github.com/open-feature/dotnet-sdk-contrib/issues/478
         Skip.If(tags.Contains("sync-port"), "Skipping sync-port as it is not supported.");
+=======
+        Skip.If(!tags.Contains("in-process"), "Skipping scenario because it does not have required tag.");
+        Skip.If(tags.Contains("fractional-v1"), "Skipping legacy fractional bucketing test; v2 algorithm is implemented.");
+        Skip.If(tags.Contains("operator-errors"), "Skipping operator-errors test; flagd server does not yet fall back to default on operator errors.");
+>>>>>>> main
     }
 }